# scx_lavd: Fix CPU topology sorting for ARM big.LITTLE systems