## Description
Fix CPU topology sorting in performance mode for ARM big.LITTLE systems with 3 cluster types.

## Problem
On ARM big.LITTLE systems with shared LLC (like Google Pixel 6), the original sorting logic prioritized physical CPU position (`cpu_pos`) over performance capacity (`cpu_cap`), causing latency-critical tasks to be scheduled on low-performance cores.

### Current Behavior (Incorrect)
```
CPU pref order in performance mode: [2, 3, 0, 1, 4, 5, 6, 7]
```
- Little cores (160 capacity) get priority over prime cores (1024 capacity)

### Expected Behavior (Fixed)
```
CPU pref order in performance mode: [6, 7, 4, 5, 0, 1, 2, 3]  
```
- Prime cores (1024) → Big cores (498) → Little cores (160)

## Solution
Reorder the sorting criteria in performance mode to prioritize:
1. NUMA node locality (`node_id`)
2. LLC locality (`llc_pos`) 
3. **CPU performance capacity (`cpu_cap`)** ← Key fix
4. Physical position (`cpu_pos`) as tie-breaker

## Testing
Verified on Google Pixel 6 (Google Tensor SoC) with TikTok performance testing using scx_lavd autopilot mode:
### Performance Comparison
| Metric | Before Fix | After Fix | Improvement |
|--------|------------|-----------|-------------|
| **Average FPS** | **18.25 FPS** | **18.65 FPS** | **+2.2%** |
| Jank | 43.5-57.1 | 4.9-31.6 | **~40-85% reduction** |

## Impact
- Fixes latency-critical task scheduling on ARM big.LITTLE systems
- Improves interactive performance on mobile devices

Fixes #1991