@CosmicBulb 2025-05-26
Hi @EricccTaiwan, 

Thanks for the heads up! I've just fixed the formatting issues by running `cargo fmt` and pushed the updated commit. The CI should pass now.

Looking forward to the review from other maintainers. This fix addresses a significant performance issue on ARM big.LITTLE systems where latency-critical tasks were being incorrectly scheduled on low-performance cores.

Thanks for your help! 👍