@multics69 2025-05-26
Thank you super much @CosmicBulb for sending a PR!
I wanted to test your change on my side. To that end, could you share the following with me?
  - A tarball under `/sys/devices/system/cpu`
  - A tarball under `/sys/kernel/debug/energy_model/`
  - The output of `lscpu -e`