@CosmicBulb 2025-05-27
Hi @multics69,

Thank you for reviewing this PR! I've prepared the complete system information from my Google Pixel 6 (Google Tensor SoC) .

##  **System Information**

### lscpu -e Output
```
root@localhost:/sys/devices/system/cpu# lscpu -e
CPU SOCKET CORE L1d:L1i:L2:L3 ONLINE    MAXMHZ   MINMHZ      MHZ
  0      0    0 0:0:0:0          yes 1803.0000 300.0000 738.0000
  1      0    1 1:1:0:0          yes 1803.0000 300.0000 738.0000
  2      0    2 2:2:0:0          yes 1803.0000 300.0000 738.0000
  3      0    3 3:3:0:0          yes 1803.0000 300.0000 738.0000
  4      0    0 4:4:0:0          yes 2253.0000 400.0000 696.0000
  5      0    1 5:5:0:0          yes 2253.0000 400.0000 696.0000
  6      0    0 6:6:0:0          yes 2802.0000 500.0000 500.0000
  7      0    1 7:7:0:0          yes 2802.0000 500.0000 500.0000
```
##  **System Files**

I've attached two comprehensive archives:

### [cpu_info.tar.gz](https://github.com/user-attachments/files/20447797/cpu_info.tar.gz)
Complete CPU topology information from `/sys/devices/system/cpu/`. 

**Note**: Some symbolic links may appear broken due to tar path handling, but all essential topology data is preserved 

### [energy_model.tar.gz](https://github.com/user-attachments/files/20447798/energy_model.tar.gz)
Complete energy model data from `/sys/kernel/debug/energy_model/` with actual performance state values:

Let me know if you need any additional information or specific tests!

Thanks for your time reviewing this fix! 🙏